First iteration for login
Created layout for home page and moved header and footer to the layout from the page component.
Added netlifyIdentity widget. Initial test to open the widget works. We need to upload to the real app so that functionality would be possible on localhost as well.
Decreased z-index of original video thumbnail from reaction list since it was going over the widget